requirements.txt
cassio==0.1.5
datasets==2.18.0
langchain==0.1.11
openai==1.13.3
PyPDF2==3.0.1
tiktoken==0.6.0
fastapi
uvicorn
python-dotenv
python-multipart